+++ Cargo.toml
@@ -39,14 +39,6 @@ name = "nelder_mead_benchmark"
 harness = false
 
 [[bench]]
-name = "bfgs_benchmark"
-harness = false
-
-[[bench]]
-name = "lbfgs_benchmark"
-harness = false
-
-[[bench]]
 name = "lbfgsb_benchmark"
 harness = false
