+++ benches/nelder_mead_benchmark.rs
@@ -7,12 +7,12 @@ fn nelder_mead_benchmark(c: &mut Criterion) {
     let mut group = c.benchmark_group("Nelder Mead");
     for n in [2, 3, 4, 5] {
         group.bench_with_input(BenchmarkId::new("Rosenbrock", n), &n, |b, ndim| {
-            let mut problem = Rosenbrock { n: *ndim };
+            let problem = Rosenbrock { n: *ndim };
             let nm = NelderMead::default();
             let mut m = Minimizer::new(Box::new(nm), *ndim).with_max_steps(10_000_000);
             let x0 = vec![5.0; *ndim];
             b.iter(|| {
-                m.minimize(&mut problem, &x0, &mut ()).unwrap();
+                m.minimize(&problem, &x0, &mut ()).unwrap();
                 black_box(&m.status);
             });
         });
@@ -20,12 +20,12 @@ fn nelder_mead_benchmark(c: &mut Criterion) {
             BenchmarkId::new("Rosenbrock (adaptive)", n),
             &n,
             |b, ndim| {
-                let mut problem = Rosenbrock { n: *ndim };
+                let problem = Rosenbrock { n: *ndim };
                 let nm = NelderMead::default().with_adaptive(n);
                 let mut m = Minimizer::new(Box::new(nm), *ndim).with_max_steps(10_000_000);
                 let x0 = vec![5.0; *ndim];
                 b.iter(|| {
-                    m.minimize(&mut problem, &x0, &mut ()).unwrap();
+                    m.minimize(&problem, &x0, &mut ()).unwrap();
                     black_box(&m.status);
                 });
             },
