+++ src/algorithms/line_search.rs
@@ -1,7 +1,7 @@
 use dyn_clone::DynClone;
 use nalgebra::DVector;
 
-use crate::{Bound, Float, Function, Status};
+use crate::{Float, Function, Status};
 
 /// A trait which defines the methods for a line search algorithm.
 ///
@@ -25,7 +25,6 @@ pub trait LineSearch<U, E>: DynClone {
         p: &DVector<Float>,
         max_step: Option<Float>,
         func: &dyn Function<U, E>,
-        bounds: Option<&Vec<Bound>>,
         user_data: &mut U,
         status: &mut Status,
     ) -> Result<(bool, Float, Float, DVector<Float>), E>;
@@ -54,20 +53,17 @@ impl<U, E> LineSearch<U, E> for BacktrackingLineSearch {
         p: &DVector<Float>,
         max_step: Option<Float>,
         func: &dyn Function<U, E>,
-        bounds: Option<&Vec<Bound>>,
         user_data: &mut U,
         status: &mut Status,
     ) -> Result<(bool, Float, Float, DVector<Float>), E> {
         let mut alpha_i = max_step.map_or(1.0, |max_alpha| max_alpha);
         let phi = |alpha: Float, ud: &mut U, st: &mut Status| -> Result<Float, E> {
             st.inc_n_f_evals();
-            func.evaluate_bounded((x + p.scale(alpha)).as_slice(), bounds, ud)
+            func.evaluate((x + p.scale(alpha)).as_slice(), ud)
         };
         let dphi = |alpha: Float, ud: &mut U, st: &mut Status| -> Result<Float, E> {
             st.inc_n_g_evals();
-            Ok(func
-                .gradient_bounded((x + p.scale(alpha)).as_slice(), bounds, ud)?
-                .dot(p))
+            Ok(func.gradient((x + p.scale(alpha)).as_slice(), ud)?.dot(p))
         };
         let phi_0 = phi(0.0, user_data, status)?;
         let mut phi_alpha_i = phi(alpha_i, user_data, status)?;
@@ -75,8 +71,7 @@ impl<U, E> LineSearch<U, E> for BacktrackingLineSearch {
         loop {
             let armijo = phi_alpha_i <= (self.c * alpha_i).mul_add(dphi_0, phi_0);
             if armijo {
-                let g_alpha_i =
-                    func.gradient_bounded((x + p.scale(alpha_i)).as_slice(), bounds, user_data)?;
+                let g_alpha_i = func.gradient((x + p.scale(alpha_i)).as_slice(), user_data)?;
                 return Ok((true, alpha_i, phi_alpha_i, g_alpha_i));
             }
             alpha_i *= self.rho;
@@ -159,39 +154,27 @@ impl StrongWolfeLineSearch {
         &self,
         func: &dyn Function<U, E>,
         x: &DVector<Float>,
-        bounds: Option<&Vec<Bound>>,
         user_data: &mut U,
         status: &mut Status,
     ) -> Result<Float, E> {
         status.inc_n_f_evals();
-        if self.use_bounds {
-            func.evaluate_bounded(x.as_slice(), bounds, user_data)
-        } else {
-            func.evaluate(x.as_slice(), user_data)
-        }
+        func.evaluate(x.as_slice(), user_data)
     }
     fn g_eval<U, E>(
         &self,
         func: &dyn Function<U, E>,
         x: &DVector<Float>,
-        bounds: Option<&Vec<Bound>>,
         user_data: &mut U,
         status: &mut Status,
     ) -> Result<DVector<Float>, E> {
         status.inc_n_g_evals();
-        if self.use_bounds {
-            func.gradient_bounded(x.as_slice(), bounds, user_data)
-                .map(DVector::from)
-        } else {
-            func.gradient(x.as_slice(), user_data).map(DVector::from)
-        }
+        func.gradient(x.as_slice(), user_data).map(DVector::from)
     }
     #[allow(clippy::too_many_arguments)]
     fn zoom<U, E>(
         &self,
         func: &dyn Function<U, E>,
         x0: &DVector<Float>,
-        bounds: Option<&Vec<Bound>>,
         user_data: &mut U,
         f0: Float,
         g0: &DVector<Float>,
@@ -207,14 +190,14 @@ impl StrongWolfeLineSearch {
         loop {
             let alpha_i = (alpha_lo + alpha_hi) / 2.0;
             let x = x0 + p.scale(alpha_i);
-            let f_i = self.f_eval(func, &x, bounds, user_data, status)?;
+            let f_i = self.f_eval(func, &x, user_data, status)?;
             let x_lo = x0 + p.scale(alpha_lo);
-            let f_lo = self.f_eval(func, &x_lo, bounds, user_data, status)?;
+            let f_lo = self.f_eval(func, &x_lo, user_data, status)?;
             let valid = if (f_i > (self.c1 * alpha_i).mul_add(dphi0, f0)) || (f_i >= f_lo) {
                 alpha_hi = alpha_i;
                 false
             } else {
-                let g_i = self.g_eval(func, &x, bounds, user_data, status)?;
+                let g_i = self.g_eval(func, &x, user_data, status)?;
                 let dphi = g_i.dot(p);
                 if Float::abs(dphi) <= -self.c2 * dphi0 {
                     return Ok((true, alpha_i, f_i, g_i));
@@ -227,7 +210,7 @@ impl StrongWolfeLineSearch {
             };
             i += 1;
             if i > self.max_zoom {
-                let g_i = self.g_eval(func, &x, bounds, user_data, status)?;
+                let g_i = self.g_eval(func, &x, user_data, status)?;
                 return Ok((valid, alpha_i, f_i, g_i));
             }
         }
@@ -241,12 +224,11 @@ impl<U, E> LineSearch<U, E> for StrongWolfeLineSearch {
         p: &DVector<Float>,
         max_step: Option<Float>,
         func: &dyn Function<U, E>,
-        bounds: Option<&Vec<Bound>>,
         user_data: &mut U,
         status: &mut Status,
     ) -> Result<(bool, Float, Float, DVector<Float>), E> {
-        let f0 = self.f_eval(func, x0, bounds, user_data, status)?;
-        let g0 = self.g_eval(func, x0, bounds, user_data, status)?;
+        let f0 = self.f_eval(func, x0, user_data, status)?;
+        let g0 = self.g_eval(func, x0, user_data, status)?;
         let alpha_max = max_step.map_or(1.0, |alpha_max| alpha_max);
         let mut alpha_im1 = 0.0;
         let mut alpha_i = 1.0;
@@ -255,21 +237,17 @@ impl<U, E> LineSearch<U, E> for StrongWolfeLineSearch {
         let mut i = 0;
         loop {
             let x = x0 + p.scale(alpha_i);
-            let f_i = self.f_eval(func, &x, bounds, user_data, status)?;
+            let f_i = self.f_eval(func, &x, user_data, status)?;
             if (f_i > self.c1.mul_add(dphi0, f0)) || (i > 1 && f_i >= f_im1) {
-                return self.zoom(
-                    func, x0, bounds, user_data, f0, &g0, p, alpha_im1, alpha_i, status,
-                );
+                return self.zoom(func, x0, user_data, f0, &g0, p, alpha_im1, alpha_i, status);
             }
-            let g_i = self.g_eval(func, &x, bounds, user_data, status)?;
+            let g_i = self.g_eval(func, &x, user_data, status)?;
             let dphi = g_i.dot(p);
             if Float::abs(dphi) <= self.c2 * Float::abs(dphi0) {
                 return Ok((true, alpha_i, f_i, g_i));
             }
             if dphi >= 0.0 {
-                return self.zoom(
-                    func, x0, bounds, user_data, f0, &g0, p, alpha_i, alpha_im1, status,
-                );
+                return self.zoom(func, x0, user_data, f0, &g0, p, alpha_i, alpha_im1, status);
             }
             alpha_im1 = alpha_i;
             f_im1 = f_i;
