+++ src/algorithms/nelder_mead.rs
@@ -2,7 +2,7 @@ use std::fmt::Debug;
 
 use nalgebra::{DMatrix, DVector};
 
-use crate::{algorithms::Point, Algorithm, Bound, Float, Function, Status};
+use crate::{Algorithm, Bound, Float, Function, Point, Status};
 
 /// Gives a method for constructing a simplex.
 #[derive(Debug, Clone)]
@@ -30,14 +30,13 @@ impl SimplexConstructionMethod {
         &self,
         func: &dyn Function<U, E>,
         x0: &[Float],
-        bounds: Option<&Vec<Bound>>,
         user_data: &mut U,
     ) -> Result<Simplex, E> {
         match self {
             Self::Orthogonal { simplex_size } => {
                 let mut points = Vec::default();
-                let mut point_0 = Point::from(Bound::to_unbounded(x0, bounds));
-                point_0.evaluate_bounded(func, bounds, user_data)?;
+                let mut point_0 = Point::from(Bound::to_unbounded(x0, func.bounds()));
+                point_0.evaluate(func, user_data)?;
                 points.push(point_0.clone());
                 let dim = point_0.dimension();
                 assert!(
@@ -48,7 +47,7 @@ impl SimplexConstructionMethod {
                     let mut point_i = point_0.clone();
                     point_i.x[i] += *simplex_size;
                     point_i.fx = Float::NAN;
-                    point_i.evaluate_bounded(func, bounds, user_data)?;
+                    point_i.evaluate(func, user_data)?;
                     points.push(point_i);
                 }
                 Ok(Simplex::new(&points))
@@ -61,8 +60,8 @@ impl SimplexConstructionMethod {
                     &simplex
                         .iter()
                         .map(|x| {
-                            let mut point_i = Point::from(Bound::to_unbounded(x, bounds));
-                            point_i.evaluate_bounded(func, bounds, user_data)?;
+                            let mut point_i = Point::from(Bound::to_unbounded(x, func.bounds()));
+                            point_i.evaluate(func, user_data)?;
                             Ok(point_i)
                         })
                         .collect::<Result<Vec<Point>, E>>()?,
@@ -201,47 +200,44 @@ pub enum NelderMeadFTerminator {
     /// ```math
     /// 2 \frac{f(x_h) - f(x_l)}{|f(x_h)| + |f(x_l)|} <= \varepsilon
     /// ```
-    Amoeba {
-        /// Relative tolerance $`\varepsilon`$.
-        tol_f_rel: Float,
-    },
+    Amoeba,
     /// For the worst point $`x_h`$ and best point $`x_l`$, converge if the following is true:
     /// ```math
     /// f(x_h) - f(x_l) <= \varepsilon
     /// ```
-    Absolute {
-        /// Absolute tolerance $`\varepsilon`$.
-        tol_f_abs: Float,
-    },
+    Absolute,
     /// Converge if the standard deviation of the function evaluations of all points in the simplex
     /// is $`\sigma <= \varepsilon`$.
-    StdDev {
-        /// Absolute tolerance $`\varepsilon`$.
-        tol_f_abs: Float,
-    },
+    StdDev,
     /// No termination condition.
     None,
 }
 impl NelderMeadFTerminator {
-    fn update_convergence(&self, simplex: &Simplex, status: &mut Status) {
+    fn update_convergence(
+        &self,
+        simplex: &Simplex,
+        status: &mut Status,
+        eps_rel: Float,
+        eps_abs: Float,
+    ) {
         match self {
-            Self::Amoeba { tol_f_rel } => {
+            Self::Amoeba => {
                 let fh = simplex.worst().get_fx_checked();
                 let fl = simplex.best().get_fx_checked();
-                if 2.0 * (fh - fl) / (Float::abs(fh) + Float::abs(fl)) <= *tol_f_rel {
+                if 2.0 * (fh - fl) / (Float::abs(fh) + Float::abs(fl)) <= eps_rel {
                     status.set_converged();
                     status.update_message("term_f = AMOEBA");
                 }
             }
-            Self::Absolute { tol_f_abs } => {
+            Self::Absolute => {
                 let fh = simplex.worst().get_fx_checked();
                 let fl = simplex.best().get_fx_checked();
-                if fh - fl <= *tol_f_abs {
+                if fh - fl <= eps_abs {
                     status.set_converged();
                     status.update_message("term_f = ABSOLUTE");
                 }
             }
-            Self::StdDev { tol_f_abs } => {
+            Self::StdDev => {
                 let dim = simplex.dimension as Float;
                 let mean = simplex.points.iter().map(|point| point.fx).sum::<Float>() / dim;
                 let std_dev = Float::sqrt(
@@ -252,7 +248,7 @@ impl NelderMeadFTerminator {
                         .sum::<Float>()
                         / dim,
                 );
-                if std_dev <= *tol_f_abs {
+                if std_dev <= eps_abs {
                     status.set_converged();
                     status.update_message("term_f = STDDEV")
                 }
@@ -272,28 +268,19 @@ pub enum NelderMeadXTerminator {
     /// ```math
     /// \max_{j\neq l} ||x_j - x_l||_{\inf} \leq \varepsilon
     /// ```
-    Diameter {
-        /// Absolute tolerance $`\varepsilon`$.
-        tol_x_abs: Float,
-    },
+    Diameter,
     /// For the best point in the simplex $`x_l`$, converge if the following condition is met:
     /// ```math
     /// \frac{\max_{j\neq l} ||x_j - x_l||_1}{\max\left\{1, ||x_l||_1\right\}} \leq \varepsilon
     /// ```
-    Higham {
-        /// Relative tolerance $`\varepsilon`$.
-        tol_x_rel: Float,
-    },
+    Higham,
     /// For the worst point $`x_h`$ and best point $`x_l`$, as well as the original values of those
     /// points at the beginning of the algorithm, denoted $`x_h^{(0)}`$ and $`x_l^{(0)}`$
     /// respectively, converge if the following condition is met:
     /// ```math
     /// ||x_h - x_l||_2 \leq \varepsilon ||x_h^{(0)} - x_l^{(0)}||_2
     /// ```
-    Rowan {
-        /// Relative tolerance $`\varepsilon`$.
-        tol_x_rel: Float,
-    },
+    Rowan,
     /// Given the volume of the simplex
     /// ```math
     /// V(S) \equiv \frac{1}{n!}\sqrt{\Gamma(x_1-x_0,...,x_n-x_0)}
@@ -306,18 +293,21 @@ pub enum NelderMeadXTerminator {
     /// where $`S`$ is the current simplex and $`S^{(0)}`$ is the simplex at the beginning of the
     /// algorithm. Note that $`V(S)`$ is only calculated once in practice and updated according to
     /// each step type by a single multiplication, so this method is very efficient.
-    Singer {
-        /// Relative tolerance $`\varepsilon`$.
-        tol_x_rel: Float,
-    },
+    Singer,
     /// No termination condition.
     None,
 }
 
 impl NelderMeadXTerminator {
-    fn update_convergence(&self, simplex: &Simplex, status: &mut Status) {
+    fn update_convergence(
+        &self,
+        simplex: &Simplex,
+        status: &mut Status,
+        eps_rel: Float,
+        eps_abs: Float,
+    ) {
         match self {
-            Self::Diameter { tol_x_abs } => {
+            Self::Diameter => {
                 let l = simplex.worst();
                 let max_inf_norm = simplex
                     .points
@@ -336,12 +326,12 @@ impl NelderMeadXTerminator {
                     })
                     .max_by(|&a, &b| a.total_cmp(&b))
                     .unwrap_or(0.0);
-                if max_inf_norm <= *tol_x_abs {
+                if max_inf_norm <= eps_abs {
                     status.set_converged();
                     status.update_message("term_x = DIAMETER");
                 }
             }
-            Self::Higham { tol_x_rel } => {
+            Self::Higham => {
                 let l = simplex.worst();
                 let l1_norm_l = l.x.lp_norm(1);
                 let denom = Float::max(l1_norm_l, 1.0);
@@ -356,24 +346,24 @@ impl NelderMeadXTerminator {
                     })
                     .max_by(|&a, &b| a.total_cmp(&b))
                     .unwrap_or(0.0);
-                if numer / denom <= *tol_x_rel {
+                if numer / denom <= eps_rel {
                     status.set_converged();
                     status.update_message("term_x = HIGHAM");
                 }
             }
-            Self::Rowan { tol_x_rel } => {
+            Self::Rowan => {
                 let init_diff = (&simplex.initial_worst.x - &simplex.initial_best.x).lp_norm(2);
                 let current_diff = (&simplex.worst().x - &simplex.best().x).lp_norm(2);
-                if current_diff <= *tol_x_rel * init_diff {
+                if current_diff <= eps_rel * init_diff {
                     status.set_converged();
                     status.update_message("term_x = ROWAN");
                 }
             }
-            Self::Singer { tol_x_rel } => {
+            Self::Singer => {
                 let dim = simplex.dimension as Float;
                 let lv_init = Float::powf(simplex.initial_volume, 1.0 / dim);
                 let lv_current = Float::powf(simplex.volume, 1.0 / dim);
-                if lv_current <= *tol_x_rel * lv_init {
+                if lv_current <= eps_rel * lv_init {
                     status.set_converged();
                     status.update_message("term_x = SINGER");
                 }
@@ -425,6 +415,10 @@ pub struct NelderMead {
     terminator_f: NelderMeadFTerminator,
     terminator_x: NelderMeadXTerminator,
     compute_parameter_errors: bool,
+    eps_x_rel: Float,
+    eps_x_abs: Float,
+    eps_f_rel: Float,
+    eps_f_abs: Float,
 }
 impl Default for NelderMead {
     fn default() -> Self {
@@ -443,14 +437,54 @@ impl NelderMead {
             simplex: Simplex::default(),
             construction_method: SimplexConstructionMethod::default(),
             expansion_method: SimplexExpansionMethod::default(),
-            terminator_f: NelderMeadFTerminator::StdDev {
-                tol_f_abs: Float::EPSILON.powf(0.25),
-            },
-            terminator_x: NelderMeadXTerminator::Singer {
-                tol_x_rel: Float::EPSILON.powf(0.25),
-            },
+            terminator_f: NelderMeadFTerminator::StdDev,
+            terminator_x: NelderMeadXTerminator::Singer,
             compute_parameter_errors: true,
+            eps_x_rel: Float::EPSILON.powf(0.25),
+            eps_x_abs: Float::EPSILON.powf(0.25),
+            eps_f_rel: Float::EPSILON.powf(0.25),
+            eps_f_abs: Float::EPSILON.powf(0.25),
         }
+    }
+    /// Set the relative x-convergence tolerance (default = `MACH_EPS^(1/4)`).
+    ///
+    /// # Panics
+    ///
+    /// This method will panic if $`\epsilon <= 0`$.
+    pub fn with_eps_x_rel(mut self, value: Float) -> Self {
+        assert!(value > 0.0);
+        self.eps_x_rel = value;
+        self
+    }
+    /// Set the absolute x-convergence tolerance (default = `MACH_EPS^(1/4)`).
+    ///
+    /// # Panics
+    ///
+    /// This method will panic if $`\epsilon <= 0`$.
+    pub fn with_eps_x_abs(mut self, value: Float) -> Self {
+        assert!(value > 0.0);
+        self.eps_x_abs = value;
+        self
+    }
+    /// Set the relative f-convergence tolerance (default = `MACH_EPS^(1/4)`).
+    ///
+    /// # Panics
+    ///
+    /// This method will panic if $`\epsilon <= 0`$.
+    pub fn with_eps_f_rel(mut self, value: Float) -> Self {
+        assert!(value > 0.0);
+        self.eps_f_rel = value;
+        self
+    }
+    /// Set the absolute f-convergence tolerance (default = `MACH_EPS^(1/4)`).
+    ///
+    /// # Panics
+    ///
+    /// This method will panic if $`\epsilon <= 0`$.
+    pub fn with_eps_f_abs(mut self, value: Float) -> Self {
+        assert!(value > 0.0);
+        self.eps_f_abs = value;
+        self
     }
     /// Set the reflection coefficient $`\alpha`$ (default = `1`).
     ///
@@ -546,14 +580,11 @@ impl<U, E> Algorithm<U, E> for NelderMead {
         &mut self,
         func: &dyn Function<U, E>,
         x0: &[Float],
-        bounds: Option<&Vec<Bound>>,
         user_data: &mut U,
         status: &mut Status,
     ) -> Result<(), E> {
-        self.simplex = self
-            .construction_method
-            .generate(func, x0, bounds, user_data)?;
-        status.update_position(self.simplex.best_position(bounds));
+        self.simplex = self.construction_method.generate(func, x0, user_data)?;
+        status.update_position(self.simplex.best_position(func.bounds()));
         Ok(())
     }
 
@@ -561,7 +592,6 @@ impl<U, E> Algorithm<U, E> for NelderMead {
         &mut self,
         _i_step: usize,
         func: &dyn Function<U, E>,
-        bounds: Option<&Vec<Bound>>,
         user_data: &mut U,
         status: &mut Status,
     ) -> Result<(), E> {
@@ -570,7 +600,7 @@ impl<U, E> Algorithm<U, E> for NelderMead {
         let l = self.simplex.best();
         let c = &self.simplex.centroid;
         let mut xr = Point::from(c + (c - &h.x).scale(self.alpha));
-        xr.evaluate_bounded(func, bounds, user_data)?;
+        xr.evaluate(func, user_data)?;
         status.inc_n_f_evals();
         if l <= &xr && &xr < s {
             // Reflect if l <= x_r < s
@@ -578,7 +608,7 @@ impl<U, E> Algorithm<U, E> for NelderMead {
             // it should go. We have to do a sort, but it should be quick since most of the simplex
             // is already sorted.
             self.simplex.insert_and_sort(self.simplex.dimension - 2, xr);
-            status.update_position(self.simplex.best_position(bounds));
+            status.update_position(self.simplex.best_position(func.bounds()));
             status.update_message("REFLECT");
             self.simplex.scale_volume(self.alpha);
             return Ok(());
@@ -588,7 +618,7 @@ impl<U, E> Algorithm<U, E> for NelderMead {
             // accept the expanded point x_e regardless (greedy expansion), or we should do one
             // final comparison between x_r and x_e and choose the smallest (greedy minimization).
             let mut xe = Point::from(c + (&xr.x - c).scale(self.beta));
-            xe.evaluate_bounded(func, bounds, user_data)?;
+            xe.evaluate(func, user_data)?;
             status.inc_n_f_evals();
             self.simplex.insert_sorted(
                 0,
@@ -603,7 +633,7 @@ impl<U, E> Algorithm<U, E> for NelderMead {
                     SimplexExpansionMethod::GreedyExpansion => xe,
                 },
             );
-            status.update_position(self.simplex.best_position(bounds));
+            status.update_position(self.simplex.best_position(func.bounds()));
             status.update_message("EXPAND");
             self.simplex.scale_volume(self.alpha * self.beta);
             return Ok(());
@@ -625,14 +655,14 @@ impl<U, E> Algorithm<U, E> for NelderMead {
             if &xr < h {
                 // Try to contract outside if x_r < h
                 let mut xc = Point::from(c + (&xr.x - c).scale(self.gamma));
-                xc.evaluate_bounded(func, bounds, user_data)?;
+                xc.evaluate(func, user_data)?;
                 status.inc_n_f_evals();
                 if xc <= xr {
                     if &xc < s {
                         // If we are better than the second-worst, we need to sort everything, we
                         // could technically be anywhere, even in a new best.
                         self.simplex.insert_and_sort(self.simplex.dimension - 1, xc);
-                        status.update_position(self.simplex.best_position(bounds));
+                        status.update_position(self.simplex.best_position(func.bounds()));
                     } else {
                         // Otherwise, we don't even need to update the best position, this was just
                         // a new worst or equal to second worst.
@@ -646,14 +676,14 @@ impl<U, E> Algorithm<U, E> for NelderMead {
             } else {
                 // Contract inside if h <= x_r
                 let mut xc = Point::from(c + (&h.x - c).scale(self.gamma));
-                xc.evaluate_bounded(func, bounds, user_data)?;
+                xc.evaluate(func, user_data)?;
                 status.inc_n_f_evals();
                 if &xc < h {
                     if &xc < s {
                         // If we are better than the second-worst, we need to sort everything, we
                         // could technically be anywhere, even in a new best.
                         self.simplex.insert_and_sort(self.simplex.dimension - 1, xc);
-                        status.update_position(self.simplex.best_position(bounds));
+                        status.update_position(self.simplex.best_position(func.bounds()));
                     } else {
                         // Otherwise, we don't even need to update the best position, this was just
                         // a new worst or equal to second worst.
@@ -669,7 +699,7 @@ impl<U, E> Algorithm<U, E> for NelderMead {
         let l_clone = l.clone();
         for p in self.simplex.points.iter_mut().skip(1) {
             *p = Point::from(&l_clone.x + (&p.x - &l_clone.x).scale(self.delta));
-            p.evaluate_bounded(func, bounds, user_data)?;
+            p.evaluate(func, user_data)?;
             status.inc_n_f_evals();
         }
         // We must do a fresh sort here, since we don't know the ordering of the shrunken simplex,
@@ -678,7 +708,7 @@ impl<U, E> Algorithm<U, E> for NelderMead {
         self.simplex.sort();
         // We also need to recalculate the centroid and figure out if there's a new best position:
         self.simplex.compute_centroid();
-        status.update_position(self.simplex.best_position(bounds));
+        status.update_position(self.simplex.best_position(func.bounds()));
         status.update_message("SHRINK");
         self.simplex
             .scale_volume(Float::powi(self.delta, self.simplex.dimension as i32));
@@ -688,15 +718,16 @@ impl<U, E> Algorithm<U, E> for NelderMead {
     fn check_for_termination(
         &mut self,
         _func: &dyn Function<U, E>,
-        _bounds: Option<&Vec<Bound>>,
         _user_data: &mut U,
         status: &mut Status,
     ) -> Result<bool, E> {
-        self.terminator_x.update_convergence(&self.simplex, status);
+        self.terminator_x
+            .update_convergence(&self.simplex, status, self.eps_x_rel, self.eps_x_abs);
         if status.converged {
             return Ok(true);
         }
-        self.terminator_f.update_convergence(&self.simplex, status);
+        self.terminator_f
+            .update_convergence(&self.simplex, status, self.eps_f_rel, self.eps_f_abs);
         if status.converged {
             return Ok(true);
         }
@@ -706,7 +737,6 @@ impl<U, E> Algorithm<U, E> for NelderMead {
     fn postprocessing(
         &mut self,
         func: &dyn Function<U, E>,
-        _bounds: Option<&Vec<Bound>>,
         user_data: &mut U,
         status: &mut Status,
     ) -> Result<(), E> {
@@ -724,7 +754,7 @@ mod tests {
 
     use approx::assert_relative_eq;
 
-    use crate::{test_functions::Rosenbrock, Float, Minimizer};
+    use crate::{test_functions::Rosenbrock, BoundedFunction, Float, Minimizer};
 
     use super::NelderMead;
 
@@ -732,23 +762,23 @@ mod tests {
     fn test_nelder_mead() -> Result<(), Infallible> {
         let algo = NelderMead::default();
         let mut m = Minimizer::new(Box::new(algo), 2);
-        let mut problem = Rosenbrock { n: 2 };
-        m.minimize(&mut problem, &[-2.0, 2.0], &mut ())?;
+        let problem = Rosenbrock { n: 2 };
+        m.minimize(&problem, &[-2.0, 2.0], &mut ())?;
         assert!(m.status.converged);
         assert_relative_eq!(m.status.fx, 0.0, epsilon = Float::EPSILON.powf(0.25));
-        m.minimize(&mut problem, &[2.0, 2.0], &mut ())?;
+        m.minimize(&problem, &[2.0, 2.0], &mut ())?;
         assert!(m.status.converged);
         assert_relative_eq!(m.status.fx, 0.0, epsilon = Float::EPSILON.powf(1.0 / 5.0));
-        m.minimize(&mut problem, &[2.0, -2.0], &mut ())?;
+        m.minimize(&problem, &[2.0, -2.0], &mut ())?;
         assert!(m.status.converged);
         assert_relative_eq!(m.status.fx, 0.0, epsilon = Float::EPSILON.powf(0.25));
-        m.minimize(&mut problem, &[-2.0, -2.0], &mut ())?;
+        m.minimize(&problem, &[-2.0, -2.0], &mut ())?;
         assert!(m.status.converged);
         assert_relative_eq!(m.status.fx, 0.0, epsilon = Float::EPSILON.powf(0.25));
-        m.minimize(&mut problem, &[0.0, 0.0], &mut ())?;
+        m.minimize(&problem, &[0.0, 0.0], &mut ())?;
         assert!(m.status.converged);
         assert_relative_eq!(m.status.fx, 0.0, epsilon = Float::EPSILON.powf(0.25));
-        m.minimize(&mut problem, &[1.0, 1.0], &mut ())?;
+        m.minimize(&problem, &[1.0, 1.0], &mut ())?;
         assert!(m.status.converged);
         assert_relative_eq!(m.status.fx, 0.0, epsilon = Float::EPSILON.sqrt());
         Ok(())
@@ -757,25 +787,24 @@ mod tests {
     #[test]
     fn test_bounded_nelder_mead() -> Result<(), Infallible> {
         let algo = NelderMead::default();
-        let mut m =
-            Minimizer::new(Box::new(algo), 2).with_bounds(Some(vec![(-4.0, 4.0), (-4.0, 4.0)]));
-        let mut problem = Rosenbrock { n: 2 };
-        m.minimize(&mut problem, &[-2.0, 2.0], &mut ())?;
+        let mut m = Minimizer::new(Box::new(algo), 2);
+        let problem = BoundedFunction::new(Rosenbrock { n: 2 }, vec![(-4.0, 4.0), (-4.0, 4.0)]);
+        m.minimize(&problem, &[-2.0, 2.0], &mut ())?;
         assert!(m.status.converged);
         assert_relative_eq!(m.status.fx, 0.0, epsilon = Float::EPSILON.powf(0.25));
-        m.minimize(&mut problem, &[2.0, 2.0], &mut ())?;
+        m.minimize(&problem, &[2.0, 2.0], &mut ())?;
         assert!(m.status.converged);
         assert_relative_eq!(m.status.fx, 0.0, epsilon = Float::EPSILON.powf(1.0 / 5.0));
-        m.minimize(&mut problem, &[2.0, -2.0], &mut ())?;
+        m.minimize(&problem, &[2.0, -2.0], &mut ())?;
         assert!(m.status.converged);
         assert_relative_eq!(m.status.fx, 0.0, epsilon = Float::EPSILON.powf(0.25));
-        m.minimize(&mut problem, &[-2.0, -2.0], &mut ())?;
+        m.minimize(&problem, &[-2.0, -2.0], &mut ())?;
         assert!(m.status.converged);
         assert_relative_eq!(m.status.fx, 0.0, epsilon = Float::EPSILON.powf(1.0 / 5.0));
-        m.minimize(&mut problem, &[0.0, 0.0], &mut ())?;
+        m.minimize(&problem, &[0.0, 0.0], &mut ())?;
         assert!(m.status.converged);
         assert_relative_eq!(m.status.fx, 0.0, epsilon = Float::EPSILON.powf(1.0 / 5.0));
-        m.minimize(&mut problem, &[1.0, 1.0], &mut ())?;
+        m.minimize(&problem, &[1.0, 1.0], &mut ())?;
         assert!(m.status.converged);
         assert_relative_eq!(m.status.fx, 0.0, epsilon = Float::EPSILON.sqrt());
         Ok(())
@@ -785,23 +814,23 @@ mod tests {
     fn test_adaptive_nelder_mead() -> Result<(), Infallible> {
         let algo = NelderMead::default().with_adaptive(2);
         let mut m = Minimizer::new(Box::new(algo), 2);
-        let mut problem = Rosenbrock { n: 2 };
-        m.minimize(&mut problem, &[-2.0, 2.0], &mut ())?;
+        let problem = Rosenbrock { n: 2 };
+        m.minimize(&problem, &[-2.0, 2.0], &mut ())?;
         assert!(m.status.converged);
         assert_relative_eq!(m.status.fx, 0.0, epsilon = Float::EPSILON.powf(0.25));
-        m.minimize(&mut problem, &[2.0, 2.0], &mut ())?;
+        m.minimize(&problem, &[2.0, 2.0], &mut ())?;
         assert!(m.status.converged);
         assert_relative_eq!(m.status.fx, 0.0, epsilon = Float::EPSILON.powf(1.0 / 5.0));
-        m.minimize(&mut problem, &[2.0, -2.0], &mut ())?;
+        m.minimize(&problem, &[2.0, -2.0], &mut ())?;
         assert!(m.status.converged);
         assert_relative_eq!(m.status.fx, 0.0, epsilon = Float::EPSILON.powf(0.25));
-        m.minimize(&mut problem, &[-2.0, -2.0], &mut ())?;
+        m.minimize(&problem, &[-2.0, -2.0], &mut ())?;
         assert!(m.status.converged);
         assert_relative_eq!(m.status.fx, 0.0, epsilon = Float::EPSILON.powf(0.25));
-        m.minimize(&mut problem, &[0.0, 0.0], &mut ())?;
+        m.minimize(&problem, &[0.0, 0.0], &mut ())?;
         assert!(m.status.converged);
         assert_relative_eq!(m.status.fx, 0.0, epsilon = Float::EPSILON.powf(0.25));
-        m.minimize(&mut problem, &[1.0, 1.0], &mut ())?;
+        m.minimize(&problem, &[1.0, 1.0], &mut ())?;
         assert!(m.status.converged);
         assert_relative_eq!(m.status.fx, 0.0, epsilon = Float::EPSILON.sqrt());
         Ok(())
